+++ src/core/server/events/events.ts
@@ -9,17 +9,7 @@ import {
 } from "coral-server/graph/resolvers/Subscription";
 
 import { CoralEventPayload, createCoralEvent } from "./event";
-
-export enum CoralEventType {
-  COMMENT_ENTERED_MODERATION_QUEUE = "COMMENT_ENTERED_MODERATION_QUEUE",
-  COMMENT_LEFT_MODERATION_QUEUE = "COMMENT_LEFT_MODERATION_QUEUE",
-  COMMENT_STATUS_UPDATED = "COMMENT_STATUS_UPDATED",
-  COMMENT_REPLY_CREATED = "COMMENT_REPLY_CREATED",
-  COMMENT_CREATED = "COMMENT_CREATED",
-  COMMENT_FEATURED = "COMMENT_FEATURED",
-  COMMENT_RELEASED = "COMMENT_RELEASED",
-  STORY_CREATED = "STORY_CREATED",
-}
+import { CoralEventType } from "./types";
 
 export type CommentEnteredModerationQueueCoralEventPayload = CoralEventPayload<
   CoralEventType.COMMENT_ENTERED_MODERATION_QUEUE,
