+++ src/core/server/events/listeners/subscription.ts
@@ -9,9 +9,9 @@ import {
   CommentReleasedCoralEventPayload,
   CommentReplyCreatedCoralEventPayload,
   CommentStatusUpdatedCoralEventPayload,
-  CoralEventType,
 } from "../events";
 import { CoralEventListener, CoralEventPublisherFactory } from "../publisher";
+import { CoralEventType } from "../types";
 
 type SubscriptionCoralEventListenerPayloads =
   | CommentEnteredModerationQueueCoralEventPayload
