+++ src/core/server/services/notifications/categories/reply.ts
@@ -2,7 +2,7 @@ import {
   CommentReplyCreatedCoralEventPayload,
   CommentStatusUpdatedCoralEventPayload,
   CoralEventType,
-} from "coral-server/events/events";
+} from "coral-server/events";
 import { hasPublishedStatus } from "coral-server/models/comment";
 import { getStoryTitle, getURLWithCommentID } from "coral-server/models/story";
